requirements.txt
asyncpg
aiogram
APScheduler
python-decouple
pandas